Refactor the category parser
